fix(host): migrate legacy connections before boot
