+++ crates/os-host/Cargo.toml
@@ -32,6 +32,7 @@ serde.workspace = true
 serde_json.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+toml = "1.1"
 uuid.workspace = true
 
 [dev-dependencies]
