+++ crates/os-host/src/lib.rs
@@ -6,10 +6,14 @@
 
 mod composition;
 pub mod host_mounts;
+mod legacy_connections;
 mod local;
 pub mod paths;
 
 pub use composition::{FixedBootConfig, TEMPORARY_FIXED_COMPOSITION_SUCCESSOR};
+pub use legacy_connections::{
+    ConnectionMigrationReport, LegacyConnectionPaths, migrate_legacy_connections,
+};
 pub use local::{
     AlanOsHost, AttachedNamespace, HostEndpointPaths, HostProcessReference, HostReadiness,
     HostStatus, LocalAttachment, request_host_stop, run_host_process,
